requirements.txt
requests
Flask
openstacksdk
